flight-app-py/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-distro
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-grpc